Add an XAUTHORITY script than can fix a sudo error
X11 connection rejected because of wrong authentication.

The above error can be fixed with

export XAUTHORITY=~/.Xauthority

The xauthority.sh contians this and can be implmented for all users with:

sudo ln -s /var/www/no-ssl/xauthority.sh /etc/profile.d

I am adding this to this repository becasue I already have the mlocate.sh
here although neither are directly related to apache2 configuration
which is what the no-ssl repository was created for.